refactor: move progress to a separate file
